+++ docs/DeveloperGuide.md
@@ -22,7 +22,7 @@
 
 |Version| As a ... | I want to ... | So that I can ...|
 |--------|----------|---------------|------------------|
-|seedu.duke.Duke.v1.0|new user|see usage instructions|refer to them when I forget how to use the application|
+|seedu.busybreak.BusyBreak.v1.0|new user|see usage instructions|refer to them when I forget how to use the application|
 |v2.0|user|find a to-do item by name|locate a to-do without having to go through the entire list|
 
 ## Non-Functional Requirements
