+++ src/main/java/seedu/busybreak/Activity.java
@@ -1,4 +1,4 @@
-package seedu.duke;
+package seedu.busybreak;
 
 import java.util.logging.Logger;
 import java.util.logging.Level;
