+++ src/main/java/seedu/busybreak/BudgetPlan.java
@@ -1,4 +1,4 @@
-package seedu.duke;
+package seedu.busybreak;
 import java.util.ArrayList;
 import java.util.Locale;
 import java.util.logging.Level;
