+++ src/test/java/seedu/busybreak/BudgetPlanTest.java
@@ -1,4 +1,4 @@
-package seedu.duke;
+package seedu.busybreak;
 import org.junit.jupiter.api.Test;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
