+++ src/test/java/seedu/busybreak/BusyBreakTest.java
@@ -1,4 +1,4 @@
-package seedu.duke;
+package seedu.busybreak;
 
 import java.util.ArrayList;
 
@@ -7,51 +7,51 @@ import org.junit.jupiter.api.Test;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-class DukeTest {
+class BusyBreakTest {
     @Test
     public void addActivityDataToList() {
         //test basic adding to list
-        Duke.list = new ArrayList<>();
+        BusyBreak.list = new ArrayList<>();
         Activity a = new Activity("2025-10-09", "09:00", "visit museum", "67");
-        Duke.list.add(a);
-        assertEquals(1, Duke.list.size());
-        assertTrue(Duke.list.contains(a));
+        BusyBreak.list.add(a);
+        assertEquals(1, BusyBreak.list.size());
+        assertTrue(BusyBreak.list.contains(a));
     }
 
     @Test
     public void editActivityDataInList() {
-        Duke.list = new ArrayList<>();
+        BusyBreak.list = new ArrayList<>();
         Activity a = new Activity("2025-10-10", "11:11", "film tiktok video", "69");
         Activity b = new Activity("2025-11-11", "11:11", "farm aura", "67");
-        Duke.list.add(a);
-        Duke.list.set(0, b);
-        assertEquals(1, Duke.list.size());
-        assertTrue(Duke.list.contains(b));
+        BusyBreak.list.add(a);
+        BusyBreak.list.set(0, b);
+        assertEquals(1, BusyBreak.list.size());
+        assertTrue(BusyBreak.list.contains(b));
     }
 
     @Test
     public void deleteActivityDataFromList() {
-        Duke.list = new ArrayList<>();
+        BusyBreak.list = new ArrayList<>();
         Activity a = new Activity("2025-10-10", "11:11", "film tiktok video", "69");
         Activity b = new Activity("2025-11-11", "11:11", "farm aura", "67");
-        Duke.list.add(a);
-        Duke.list.add(b);
-        Duke.list.remove(a);
-        assertEquals(1, Duke.list.size());
-        assertTrue(Duke.list.contains(b));
+        BusyBreak.list.add(a);
+        BusyBreak.list.add(b);
+        BusyBreak.list.remove(a);
+        assertEquals(1, BusyBreak.list.size());
+        assertTrue(BusyBreak.list.contains(b));
     }
 
     @Test
     void viewSpecificDate() {
-        Duke.list = new ArrayList<>();
+        BusyBreak.list = new ArrayList<>();
         Activity b = new Activity("2025-10-09", "09:00", "visit museum", "67");
         Activity c = new Activity("2025-10-10", "14:00", "go shopping", "100");
-        Duke.list.add(b);
-        Duke.list.add(c);
+        BusyBreak.list.add(b);
+        BusyBreak.list.add(c);
 
         String targetDate = "2025-10-09";
         ArrayList<Activity> result = new ArrayList<>();
-        for (Activity a : Duke.list) {
+        for (Activity a : BusyBreak.list) {
             if (a.getDate().equals(targetDate)) {
                 result.add(a);
             }
